+++ tests/Command/GruntCommandTest.php
@@ -1,0 +1,48 @@
+<?php
+
+/*
+ * This file is part of the Moodle Plugin CI package.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Copyright (c) 2017 Blackboard Inc. (http://www.blackboard.com)
+ * License http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace Moodlerooms\MoodlePluginCI\Tests\Command;
+
+use Moodlerooms\MoodlePluginCI\Command\GruntCommand;
+use Moodlerooms\MoodlePluginCI\Tests\Fake\Bridge\DummyMoodle;
+use Moodlerooms\MoodlePluginCI\Tests\Fake\Process\DummyExecute;
+use Moodlerooms\MoodlePluginCI\Tests\MoodleTestCase;
+use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Tester\CommandTester;
+
+class GruntCommandTest extends MoodleTestCase
+{
+    protected function executeCommand()
+    {
+        $command          = new GruntCommand();
+        $command->moodle  = new DummyMoodle($this->moodleDir);
+        $command->execute = new DummyExecute();
+
+        $application = new Application();
+        $application->add($command);
+
+        $commandTester = new CommandTester($application->find('grunt'));
+        $commandTester->execute([
+            'plugin'   => $this->pluginDir,
+            '--moodle' => $this->moodleDir,
+            '--tasks'  => ['css'],
+        ]);
+
+        return $commandTester;
+    }
+
+    public function testExecute()
+    {
+        $commandTester = $this->executeCommand();
+        $this->assertSame(0, $commandTester->getStatusCode());
+    }
+}